refactor(guardrails): move field selection helper into guardrails
