+++ src/aiq_agent/guardrails/dynamic_field_selection.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Dynamic middleware field-selection extensions."""
+"""Field-selection extensions for guardrails middleware."""
 
 from __future__ import annotations
